Support for Erosion functor. In the lti documentation, the kernel is set separately by calling setKernel(). In lti.rb, this is done automatically when an object is created. By default the kernel is an IcityBlockKernel of size 3. A cityBlockKernel is a 2D binary kernel for which the "ones" form a diamond as follows:
0 0 1 0 0
0 1 1 1 0
1 1 1 1 1    cityBlockKernel of size 5
0 1 1 1 0
0 0 1 0 0

Also, the kernel is made to look as if it were part of the parameters (which is not the case in the ltilib C++ code).
This way, the kernel size can be modified using the same method as any other parameter. The kernel is always a IcityBlockKernel,
only the size may be changed.